backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
cryptography==43.0.3
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
